+++ src/render/md.jl
@@ -52,10 +52,14 @@ function writemd{category}(io::IO, modname, obj, ent::Entry{category},
                     string(objname, config.md_permalink ? " [¶](#$anchorname)" : ""))
     writemd(io, docs(ent))
     println(io)
-    for k in sort(collect(keys(ent.data)))
-        println_mdstyle(io, config.mdstyle_meta, "$k:", false)
-        writemd(io, Meta{k}(ent.data[k]))
-        println(io)
+    if has_output_metadata(ent, config)
+        for m in config.metadata_order
+            if haskey(ent.data, m)
+                println_mdstyle(io, config.mdstyle_meta, "$m:", false)
+                writemd(io, Meta{m}(ent.data[m]))
+                println(io)
+            end
+        end
     end
 end